pyproject.toml
[project]
name = "wf-public-profiles"
dynamic = ["version"]

[tool.poetry]
name = "wf-public-profiles"
version = "1.0.0"
description = ""
authors = ["Benjamin Jaffe-Talberg <ben.talberg@wildflowerschools.org>"]
maintainers = []
readme = "README.md"
keywords = []
repository = "https://github.com/WildflowerSchools/wf-public-profiles"
license = "MIT"
packages = [{include = "wf-public-profiles"}]
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python"
]

[tool.poetry.dependencies]
python = "^3.11"
toml = "^0.10.2"
requests = "^2.28.2"
alembic = "^1.10.2"
sqlalchemy = "^2.0.7"
pydantic = "^1.10.7"
fastapi = "^0.95.0"
uvicorn = {extras = ["standard"], version = "^0.21.1"}
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.5"
apscheduler = "^3.10.1"
python-dateutil = "^2.8.2"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
pylint = "^2.15.0"
pytest = "^7.2.0"

[tool.black]
line-length = 120
target-version = ['py311']

[tool.setuptools]
py-modules = []

[build-system]
requires = ["poetry-core", "setuptools"]
build-backend = "poetry.core.masonry.api"
